add login handler, add logout handler
