+++ src/hooks/page/logout-handler.ts
@@ -1,0 +1,7 @@
+import LocalStorage from '@/utils/local-storage';
+
+function LogOutHandler() {
+  LocalStorage.accessToken.remove();
+  LocalStorage.firebaseAuthData.remove();
+}
+export default LogOutHandler;
